+++ internal/executor/select_test.go
@@ -855,7 +855,7 @@ func TestSelectCountStarEmptyTable(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Select() error = %v", err)
 	}
-	if len(rows) != 1 || len(rows[0]) != 1 || rows[0][0] != parser.Int64Value(0) {
+	if len(rows) != 1 || len(rows[0]) != 1 || rows[0][0] != parser.IntValue(0) {
 		t.Fatalf("Select() rows = %#v, want [[0]]", rows)
 	}
 }
@@ -873,7 +873,7 @@ func TestSelectCountStarPopulatedTable(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Select() error = %v", err)
 	}
-	if len(rows) != 1 || rows[0][0] != parser.Int64Value(3) {
+	if len(rows) != 1 || rows[0][0] != parser.IntValue(3) {
 		t.Fatalf("Select() rows = %#v, want [[3]]", rows)
 	}
 }
@@ -895,7 +895,7 @@ func TestSelectCountStarWithWhere(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Select() error = %v", err)
 	}
-	if len(rows) != 1 || rows[0][0] != parser.Int64Value(2) {
+	if len(rows) != 1 || rows[0][0] != parser.IntValue(2) {
 		t.Fatalf("Select() rows = %#v, want [[2]]", rows)
 	}
 }
@@ -943,7 +943,7 @@ func TestSelectAggregateFunctionsSingleTable(t *testing.T) {
 	if len(rows) != 1 || len(rows[0]) != 5 {
 		t.Fatalf("rows = %#v, want one aggregate row", rows)
 	}
-	if rows[0][0] != parser.Int64Value(3) || rows[0][1] != parser.RealValue((1.5+2.5+3.0)/3.0) || rows[0][2] != parser.RealValue(7.0) || rows[0][3] != parser.StringValue("alpha") || rows[0][4] != parser.RealValue(3.0) {
+	if rows[0][0] != parser.IntValue(3) || rows[0][1] != parser.RealValue((1.5+2.5+3.0)/3.0) || rows[0][2] != parser.RealValue(7.0) || rows[0][3] != parser.StringValue("alpha") || rows[0][4] != parser.RealValue(3.0) {
 		t.Fatalf("rows[0] = %#v, want [3 2.333... 7 alpha 3.0]", rows[0])
 	}
 }
@@ -982,7 +982,7 @@ func TestSelectAggregateFunctionsJoin(t *testing.T) {
 	if len(rows) != 1 || len(rows[0]) != 3 {
 		t.Fatalf("rows = %#v, want one aggregate row", rows)
 	}
-	if rows[0][0] != parser.Int64Value(3) || rows[0][1] != parser.StringValue("eng") || rows[0][2] != parser.StringValue("ops") {
+	if rows[0][0] != parser.IntValue(3) || rows[0][1] != parser.StringValue("eng") || rows[0][2] != parser.StringValue("ops") {
 		t.Fatalf("rows[0] = %#v, want [3 eng ops]", rows[0])
 	}
 }
@@ -1155,7 +1155,7 @@ func TestSelectProjectionAndPredicateArithmetic(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Select() error = %v", err)
 	}
-	if len(rows) != 1 || rows[0][0] != parser.Int64Value(3) {
+	if len(rows) != 1 || rows[0][0] != parser.IntValue(3) {
 		t.Fatalf("rows = %#v, want [[3]]", rows)
 	}
 }
@@ -1496,7 +1496,7 @@ func TestSelectIndexScanBridgeAdaptsPlannerProjectionExpressions(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Select() error = %v", err)
 	}
-	if len(rows) != 2 || rows[0][0] != parser.Int64Value(11) || rows[1][0] != parser.Int64Value(13) {
+	if len(rows) != 2 || rows[0][0] != parser.IntValue(11) || rows[1][0] != parser.IntValue(13) {
 		t.Fatalf("Select() rows = %#v, want [[11] [13]]", rows)
 	}
 }
